Remove fake "ajax" module
