+++ package.json
@@ -80,7 +80,6 @@
     "webworkify-webpack": "^1.1.3"
   },
   "browser": {
-    "./js/util/ajax.js": "./js/util/browser/ajax.js",
     "./js/util/window.js": "./js/util/browser/window.js",
     "./js/util/web_worker.js": "./js/util/browser/web_worker.js"
   },
