+++ test/fixtures/style.json
@@ -1,0 +1,5 @@
+{
+  "version": 8,
+  "sources": {},
+  "layers": []
+}
